rename: study에서 applicant분리
