+++ src/main/java/com/grepp/spring/app/model/applicant/code/ApplicantState.java
@@ -1,4 +1,4 @@
-package com.grepp.spring.app.model.study.code;
+package com.grepp.spring.app.model.applicant.code;
 
 public enum ApplicantState {
     WAIT, ACCEPT, REJECT
