+++ src/main/java/com/grepp/spring/app/model/applicant/service/ApplicantService.java
@@ -1,9 +1,9 @@
-package com.grepp.spring.app.model.study.service;
+package com.grepp.spring.app.model.applicant.service;
 
 import com.grepp.spring.app.model.studyMmeber.repository.StudyMemberRepository;
-import com.grepp.spring.app.model.study.code.ApplicantState;
-import com.grepp.spring.app.model.study.entity.Applicant;
-import com.grepp.spring.app.model.study.repository.applicant.ApplicantRepository;
+import com.grepp.spring.app.model.applicant.code.ApplicantState;
+import com.grepp.spring.app.model.applicant.entity.Applicant;
+import com.grepp.spring.app.model.applicant.repository.ApplicantRepository;
 import com.grepp.spring.infra.error.exceptions.HasNotRightException;
 import com.grepp.spring.infra.error.exceptions.NotFoundException;
 import com.grepp.spring.infra.error.exceptions.NullStateException;
