+++ src/main/java/com/grepp/spring/app/model/study/service/StudyService.java
@@ -1,7 +1,8 @@
 package com.grepp.spring.app.model.study.service;
 
 import com.grepp.spring.app.model.alarm.dto.request.AlarmRequest;
-import com.grepp.spring.app.model.study.dto.request.ApplicationResultRequest;
+import com.grepp.spring.app.model.applicant.service.ApplicantService;
+import com.grepp.spring.app.model.applicant.dto.request.ApplicationResultRequest;
 import com.grepp.spring.app.model.study.dto.request.StudyCreationRequest;
 import com.grepp.spring.app.model.study.dto.request.StudySearchRequest;
 import com.grepp.spring.app.model.study.dto.request.StudyUpdateRequest;
@@ -16,7 +17,7 @@ import com.grepp.spring.app.model.member.entity.Member;
 import com.grepp.spring.app.model.studyMmeber.entity.StudyMember;
 import com.grepp.spring.app.model.member.repository.MemberRepository;
 import com.grepp.spring.app.model.studyMmeber.repository.StudyMemberRepository;
-import com.grepp.spring.app.model.study.code.ApplicantState;
+import com.grepp.spring.app.model.applicant.code.ApplicantState;
 import com.grepp.spring.app.model.study.code.DayOfWeek;
 import com.grepp.spring.app.model.study.code.GoalType;
 import com.grepp.spring.app.model.study.code.Status;
@@ -25,12 +26,12 @@ import com.grepp.spring.app.model.study.dto.response.StudyCreationResponse;
 import com.grepp.spring.app.model.study.dto.response.StudyInfoResponse;
 import com.grepp.spring.app.model.study.dto.response.StudyListResponse;
 import com.grepp.spring.app.model.study.dto.response.WeeklyGoalStatusResponse;
-import com.grepp.spring.app.model.study.entity.Applicant;
+import com.grepp.spring.app.model.applicant.entity.Applicant;
 import com.grepp.spring.app.model.studyMmeber.entity.GoalAchievement;
 import com.grepp.spring.app.model.study.entity.Study;
 import com.grepp.spring.app.model.study.entity.StudyGoal;
 import com.grepp.spring.app.model.study.dto.response.GoalsResponse;
-import com.grepp.spring.app.model.study.repository.applicant.ApplicantRepository;
+import com.grepp.spring.app.model.applicant.repository.ApplicantRepository;
 import com.grepp.spring.app.model.studyMmeber.repository.goalAchievement.GoalAchievementRepository;
 import com.grepp.spring.app.model.study.repository.goal.StudyGoalRepository;
 import com.grepp.spring.app.model.study.repository.StudyRepository;
